chore: remove obsoleted code
