+++ src/extension/options-page/DashboardComponents/ContactLine.tsx
@@ -13,8 +13,8 @@ import { makeStyles, createStyles, Theme } from '@material-ui/core/styles'
 import MoreHorizIcon from '@material-ui/icons/MoreHoriz'
 import type { Profile } from '../../../database'
 import { Avatar } from '../../../utils/components/Avatar'
-import { useModal } from '../Dialogs/Base'
-import { DashboardContactDialog, DashboardContactDeleteConfirmDialog } from '../Dialogs/Contact'
+import { useModal } from '../DashboardDialogs/Base'
+import { DashboardContactDialog, DashboardContactDeleteConfirmDialog } from '../DashboardDialogs/Contact'
 import { Skeleton } from '@material-ui/lab'
 import DashboardMenu from './DashboardMenu'
 import { useI18N } from '../../../utils/i18n-next-ui'
