+++ src/extension/options-page/DashboardComponents/Drawer.tsx
@@ -16,8 +16,8 @@ import { Link, useRouteMatch } from 'react-router-dom'
 import { useInterval } from 'react-use'
 import ChevronRightIcon from '@material-ui/icons/ChevronRight'
 import SentimentSatisfiedOutlinedIcon from '@material-ui/icons/SentimentSatisfiedOutlined'
-import { useModal } from '../Dialogs/Base'
-import { DashboardFeedbackDialog } from '../Dialogs/Feedback'
+import { useModal } from '../DashboardDialogs/Base'
+import { DashboardFeedbackDialog } from '../DashboardDialogs/Feedback'
 import { useI18N } from '../../../utils/i18n-next-ui'
 import { cloneDeep, merge } from 'lodash-es'
 import Logo from './MaskbookLogo'
