+++ src/extension/options-page/DashboardComponents/FooterLine.tsx
@@ -3,8 +3,8 @@ import { makeStyles, createStyles } from '@material-ui/core/styles'
 import { Link } from 'react-router-dom'
 import { Breadcrumbs, Theme, Typography, Link as MuiLink } from '@material-ui/core'
 import { useI18N } from '../../../utils/i18n-next-ui'
-import { DashboardAboutDialog } from '../Dialogs/About'
-import { useModal } from '../Dialogs/Base'
+import { DashboardAboutDialog } from '../DashboardDialogs/About'
+import { useModal } from '../DashboardDialogs/Base'
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
