+++ src/extension/options-page/DashboardComponents/RestoreFromQRCodeCameraBox.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import { makeStyles, createStyles, Box, FormControl, Select, MenuItem, Button } from '@material-ui/core'
 import CropFreeIcon from '@material-ui/icons/CropFree'
-import { useModal } from '../Dialogs/Base'
-import { QRCodeVideoScannerDialog } from '../Dialogs/Setup'
+import { useModal } from '../DashboardDialogs/Base'
+import { QRCodeVideoScannerDialog } from '../DashboardDialogs/Setup'
 import { PortalShadowRoot } from '../../../utils/jss/ShadowRootPortal'
 import { useStylesExtends } from '../../../components/custom-ui-helper'
 import { useVideoDevices } from '../../../utils/hooks/useVideoDevices'
