+++ src/extension/options-page/DashboardComponents/TokenListItem.tsx
@@ -16,8 +16,8 @@ import { formatBalance } from '../../../plugins/Wallet/formatter'
 import { TokenIcon } from './TokenIcon'
 import type { WalletDetails, ERC20TokenDetails } from '../../background-script/PluginService'
 import DashboardMenu from './DashboardMenu'
-import { useModal } from '../Dialogs/Base'
-import { DashboardWalletHideTokenConfirmDialog } from '../Dialogs/Wallet'
+import { useModal } from '../DashboardDialogs/Base'
+import { DashboardWalletHideTokenConfirmDialog } from '../DashboardDialogs/Wallet'
 import { useI18N } from '../../../utils/i18n-next-ui'
 import { ETH_ADDRESS } from '../../../plugins/Wallet/token'
