+++ src/extension/options-page/DashboardComponents/WalletItem.tsx
@@ -5,7 +5,7 @@ import ChevronRightIcon from '@material-ui/icons/ChevronRight'
 import classNames from 'classnames'
 import { useCopyToClipboard } from 'react-use'
 import { useI18N } from '../../../utils/i18n-next-ui'
-import { useSnackbarCallback } from '../Dialogs/Base'
+import { useSnackbarCallback } from '../DashboardDialogs/Base'
 import ActionButton from './ActionButton'
 import { ETH_ADDRESS } from '../../../plugins/Wallet/token'
 import { TokenIcon } from './TokenIcon'
