+++ src/extension/options-page/DashboardDialogs/About.tsx
