+++ src/extension/options-page/DashboardDialogs/Wallet.tsx
