+++ src/extension/options-page/DashboardRouters/Personas.tsx
@@ -5,8 +5,8 @@ import AddIcon from '@material-ui/icons/Add'
 import AddCircleIcon from '@material-ui/icons/AddCircle'
 import RestoreIcon from '@material-ui/icons/Restore'
 import PersonaCard from '../DashboardComponents/PersonaCard'
-import { DashboardPersonaCreateDialog, DashboardImportPersonaDialog } from '../Dialogs/Persona'
-import { useModal } from '../Dialogs/Base'
+import { DashboardPersonaCreateDialog, DashboardImportPersonaDialog } from '../DashboardDialogs/Persona'
+import { useModal } from '../DashboardDialogs/Base'
 import { useI18N } from '../../../utils/i18n-next-ui'
 import { merge, cloneDeep } from 'lodash-es'
 import { useMyPersonas } from '../../../components/DataSource/independent'
